python/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt --pre setup.py
#
aiohttp==3.7.4.post0
    # via api-server (setup.py)
async-timeout==3.0.1
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
chardet==4.0.0
    # via aiohttp
idna==3.3
    # via yarl
importlib-metadata==4.8
    # via api-server (setup.py)
jsonschema==3.2.0
    # via api-server (setup.py)
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
pyrsistent==0.18.0
    # via jsonschema
pyyaml==6.0
    # via api-server (setup.py)
six==1.16.0
    # via jsonschema
typing-extensions==4.0.1
    # via aiohttp
yarl==1.7.2
    # via aiohttp
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
